+++ tests/eus9-27_slime2-20.py
@@ -1,111 +1,8 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# Test suite for EusLisp SLIME
+from euslime_test_case import EuslimeTestCase
 
-## RUN ALL TESTS:
-# ./euslime_tests.py
-
-## RUN TESTS FOR EUSLISP PROGRAM
-# ./euslime_tests.py eus
-# ./euslime_tests.py irteusgl
-# ./euslime_tests.py roseus
-
-import argparse
-import pprint
-import re
-import socket
-import time
-import unittest
-
-from euslime.logger import get_logger, set_log_level
-from euslime.server import EuslimeServer
-from threading import Thread
-
-HEADER_LENGTH = 6
-REGEX_ADDR = re.compile(r' #X[0-9a-f]*')
-
-log = get_logger(__name__)
-
-class eus(unittest.TestCase):
+class eus(EuslimeTestCase):
     EUSLISP_PROGRAM = 'eus'
     EUSLISP_PROGRAM_NAME = 'eus'
-
-    @classmethod
-    def setUpClass(self):
-        self.maxDiff = None
-        self.server = EuslimeServer(('0.0.0.0', 0), program=self.EUSLISP_PROGRAM)
-        host, port = self.server.socket.getsockname()
-        Thread(target=self.server.serve_forever).start()
-        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        # self.socket.settimeout(5)
-        self.socket.connect((host, port))
-        # Wait for process to fully start
-        time.sleep(1)
-
-    @classmethod
-    def tearDownClass(self):
-        log.info("Tearing down...")
-        self.socket.shutdown(socket.SHUT_RDWR)
-        self.socket.close()
-        log.info("...DONE")
-
-    def __del__(self):
-        try:
-            self.socket.send('')
-            # Close the socket if is still alive
-            self.tearDownClass()
-        except Exception:
-            pass
-
-    def socket_recv_one(self, *options):
-        try:
-            len = self.socket.recv(HEADER_LENGTH, *options)
-            hex_len = int(len, 16)
-            return self.socket.recv(hex_len)
-        except socket.error:
-            return;
-
-    def socket_recv(self, times):
-        result = []
-        for i in range(times):
-            res = self.socket_recv_one()
-            if res == None:
-                break
-            result.append(res)
-        return tuple(result) or None
-
-    def socket_clean(self):
-        while self.socket_recv_one(socket.MSG_DONTWAIT):
-            pass
-
-    def socket_send(self, req):
-        header = '{0:06x}'.format(len(req))
-        self.socket.send(header + req)
-
-    def socket_get_response(self, req, n):
-        log.info('request: \n%s', req)
-        # self.socket_clean()
-        self.socket_send(req)
-        return self.socket_recv(n)
-
-    def assertSocket(self, req, *res):
-        response = self.socket_get_response(req, len(res))
-        log.info('expected response: \n%s', pprint.pformat(res, width=5))
-        log.info('received response: \n%s', pprint.pformat(response, width=5))
-        self.assertEqual(res, response)
-
-    def assertSocketIgnoreAddress(self, req, *res):
-        def substitute_address(lst):
-            return [REGEX_ADDR.sub(str(), msg) for msg in lst]
-        res = substitute_address(res)
-        response = self.socket_get_response(req, len(res))
-        response = substitute_address(response)
-        log.info('expected response: \n%s', pprint.pformat(res, width=5))
-        log.info('received response: \n%s', pprint.pformat(response, width=5))
-        self.assertEqual(res, response)
-
-
-    ### TESTS
 
     # CREATE-REPL
     def test_001_create_repl(self):
@@ -705,124 +602,3 @@ class eus(unittest.TestCase):
             '(:write-string "\\n" :repl-result)',
             '(:read-aborted 0 1)',
             '(:return (:ok nil) 5)')
-
-
-class irteusgl(eus):
-    EUSLISP_PROGRAM = 'irteusgl'
-    EUSLISP_PROGRAM_NAME = 'irteusgl'
-
-    def test_completions_15(self):
-        self.assertSocket(
-            '(:emacs-rex (swank-repl:listener-eval "(progn (setq c (make-coords :name \\"test\\")) (send c :name))\n") "USER" :repl-thread 31)',
-            '(:read-string 0 1)',
-            '(:write-string "\\"test\\"" :repl-result)',
-            '(:write-string "\\n" :repl-result)',
-            '(:read-aborted 0 1)',
-            '(:return (:ok nil) 31)')
-        self.assertSocket(
-            '(:emacs-rex (swank:completions-for-keyword ":pos" (quote ("send" "c" "" swank::%cursor-marker%))) "USER" :repl-thread 35)',
-            '(:return (:ok ((":pos") ":pos")) 35)')
-        self.assertSocket(
-            '(:emacs-rex (swank:completions-for-keyword ":" (quote ("send" "c" "" swank::%cursor-marker%))) "USER" :repl-thread 37)',
-            '(:return (:ok ((":dimension" ":rot" ":pos" ":x-axis" ":y-axis" ":z-axis" ":newcoords" ":replace-rot" ":replace-pos" ":replace-coords" ":copy-rot" ":copy-pos" ":copy-coords" ":coords" ":worldrot" ":worldpos" ":worldcoords" ":copy-worldcoords" ":parentcoords" ":changed" ":reset-coords" ":move-to" ":rotate-vector" ":transform-vector" ":inverse-transform-vector" ":inverse-transformation" ":transformation" ":transform" ":rotate-with-matrix" ":rotate" ":orient-with-matrix" ":orient" ":parent-vector" ":parent-orientation" ":translate" ":locate" ":scale" ":euler" ":euler-angle" ":rpy" ":rpy-angle" ":rotation-angle" ":4x4" ":create" ":init" ":axis" ":difference-position" ":difference-rotation" ":move-coords" ":inverse-rotate-vector" ":vertices" ":draw-on" ":plist" ":get" ":put" ":name" ":remprop" ":prin1" ":warning" ":error" ":slots" ":methods" ":super" ":get-val" ":set-val") ":")) 37)')
-        self.assertSocket(
-            '(:emacs-rex (swank:completions-for-keyword ":world" (quote ("send" "c" "" swank::%cursor-marker%))) "USER" :repl-thread 39)',
-            '(:return (:ok ((":worldrot" ":worldpos" ":worldcoords") ":world")) 39)')
-        self.assertSocket(
-            '(:emacs-rex (swank-repl:listener-eval "(makunbound \'c)\n") "USER" :repl-thread 45)',
-            '(:read-string 0 1)',
-            '(:write-string "t" :repl-result)',
-            '(:write-string "\\n" :repl-result)',
-            '(:read-aborted 0 1)',
-            '(:return (:ok nil) 45)')
-
-    def test_autodoc_17(self):
-        self.assertSocket(
-            '(:emacs-rex (swank-repl:listener-eval "(progn (make-irtviewer) (send *irtviewer* :name))\n") "USER" :repl-thread 78)',
-            '(:read-string 0 1)',
-            '(:write-string "\\"IRT viewer\\"" :repl-result)',
-            '(:write-string "\\n" :repl-result)',
-            '(:read-aborted 0 1)',
-            '(:return (:ok nil) 78)')
-        self.assertSocket(
-            '(:emacs-rex (swank:autodoc (quote ("send" "*irtviewer*" ":view" swank::%cursor-marker%)) :print-right-margin 100) "USER" :repl-thread 85)',
-            '(:return (:ok ("(:view)" t)) 85)')
-        self.assertSocket(
-            '(:emacs-rex (swank:autodoc (quote ("send" "*irtviewer*" ":viewtarget" "" swank::%cursor-marker%)) :print-right-margin 100) "USER" :repl-thread 87)',
-            '(:return (:ok ("(:viewtarget &optional ===> p <===)" t)) 87)')
-        self.assertSocket(
-            '(:emacs-rex (swank:autodoc (quote ("send" "*irtviewer*" ":quit" swank::%cursor-marker%)) :print-right-margin 100) "USER" :repl-thread 93)',
-            '(:return (:ok ("(:quit &rest args)" t)) 93)')
-        self.assertSocket(
-            '(:emacs-rex (swank:autodoc (quote ("send" "*irtviewer*" ":quit" "" swank::%cursor-marker%)) :print-right-margin 100) "USER" :repl-thread 94)',
-            '(:return (:ok ("(:quit &rest ===> args <===)" t)) 94)')
-        self.assertSocket(
-            '(:emacs-rex (swank-repl:listener-eval "(send *irtviewer* :quit )\n") "USER" :repl-thread 95)',
-            '(:read-string 0 1)',
-            '(:write-string ":destroyed" :repl-result)',
-            '(:write-string "\\n" :repl-result)',
-            '(:read-aborted 0 1)',
-            '(:return (:ok nil) 95)')
-
-
-class roseus(irteusgl):
-    EUSLISP_PROGRAM = 'roseus'
-    EUSLISP_PROGRAM_NAME = 'irteusgl'
-
-    # TODO: regexp match to avoid comparing time stamp
-    # def test_rosout_1(self):
-    #     self.assertSocket(
-    #         '(:emacs-rex (swank-repl:listener-eval "(ros::ros-debug \\"HERE\\")\n") "USER" :repl-thread 13)',
-    #         '(:read-string 0 1)',
-    #         '(:write-string "t" :repl-result)',
-    #         '(:write-string "\\n" :repl-result)',
-    #         '(:read-aborted 0 1)',
-    #         '(:return (:ok nil) 13)')
-
-    # def test_rosout_2(self):
-    #     self.assertSocket(
-    #         '(:emacs-rex (swank-repl:listener-eval "(ros::ros-info \\"HERE\\")\n") "USER" :repl-thread 6)',
-    #         '(:read-string 0 1)',
-    #         '(:write-string "[ INFO] [1556256025.220748554]: HERE\\n")',
-    #         '(:write-string "t" :repl-result)',
-    #         '(:write-string "\\n" :repl-result)',
-    #         '(:read-aborted 0 1)',
-    #         '(:return (:ok nil) 6)')
-
-    # def test_rosout_3(self):
-    #     self.assertSocket(
-    #         '(:emacs-rex (swank-repl:listener-eval "(ros::ros-warn \\"HERE\\")\n") "USER" :repl-thread 9)',
-    #         '(:read-string 0 1)',
-    #         '(:write-string "[ WARN] [1556256034.693622389]: HERE\\n")',
-    #         '(:write-string "t" :repl-result)',
-    #         '(:write-string "\\n" :repl-result)',
-    #         '(:read-aborted 0 1)',
-    #         '(:return (:ok nil) 9)')
-
-    # def test_rosout_4(self):
-    #     self.assertSocket(
-    #         '(:emacs-rex (swank-repl:listener-eval "(ros::ros-error \\"HERE\\")\n") "USER" :repl-thread 12)',
-    #         '(:read-string 0 1)',
-    #         '(:write-string "[ERROR] [1556256041.151062928]: HERE\\n")',
-    #         '(:write-string "t" :repl-result)',
-    #         '(:write-string "\\n" :repl-result)',
-    #         '(:read-aborted 0 1)',
-    #         '(:return (:ok nil) 12)')
-
-    # def test_rosout_5(self):
-    #     self.assertSocket(
-    #         '(:emacs-rex (swank-repl:listener-eval "(ros::ros-fatal \\"THIS\\")\n") "USER" :repl-thread 6)',
-    #         '(:read-string 0 1)',
-    #         '(:write-string "[FATAL] [1556256284.796886986]: THIS\\n")',
-    #         '(:write-string "t" :repl-result)',
-    #         '(:write-string "\\n" :repl-result)',
-    #         '(:read-aborted 0 1)',
-    #         '(:return (:ok nil) 6)')
-
-
-if __name__ == '__main__':
-    set_log_level('debug')
-    try:
-        unittest.main()
-    except KeyboardInterrupt:
-        pass
